# apps/api/src/swagger/api-schema.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/courses": {
      "get": {
        "operationId": "CoursesController_getAllCourses",
        "parameters": [
          {
            "name": "title",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "category",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "author",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "creationDateRange[0]",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "creationDateRange[1]",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "state",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "archived",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "type": "number"
            }
          },
          {
            "name": "perPage",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "schema": {
              "anyOf": [
                {
                  "const": "title",
                  "type": "string"
                },
                {
                  "const": "category",
                  "type": "string"
                },
                {
                  "const": "creationDate",
                  "type": "string"
                },
                {
                  "const": "author",
                  "type": "string"
                },
                {
                  "const": "lessonsCount",
                  "type": "string"
                },
                {
                  "const": "enrolledParticipantsCount",
                  "type": "string"
                },
                {
                  "const": "-title",
                  "type": "string"
                },
                {
                  "const": "-category",
                  "type": "string"
                },
                {
                  "const": "-creationDate",
                  "type": "string"
                },
                {
                  "const": "-author",
                  "type": "string"
                },
                {
                  "const": "-lessonsCount",
                  "type": "string"
                },
                {
                  "const": "-enrolledParticipantsCount",
                  "type": "string"
                }
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetAllCoursesResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "CoursesController_createCourse",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCourseBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateCourseResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/courses/get-student-courses": {
      "get": {
        "operationId": "CoursesController_getStudentCourses",
        "parameters": [
          {
            "name": "title",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "category",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "author",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "creationDateRange[0]",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "creationDateRange[1]",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "state",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "archived",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "type": "number"
            }
          },
          {
            "name": "perPage",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "schema": {
              "anyOf": [
                {
                  "const": "title",
                  "type": "string"
                },
                {
                  "const": "category",
                  "type": "string"
                },
                {
                  "const": "creationDate",
                  "type": "string"
                },
                {
                  "const": "author",
                  "type": "string"
                },
                {
                  "const": "lessonsCount",
                  "type": "string"
                },
                {
                  "const": "enrolledParticipantsCount",
                  "type": "string"
                },
                {
                  "const": "-title",
                  "type": "string"
                },
                {
                  "const": "-category",
                  "type": "string"
                },
                {
                  "const": "-creationDate",
                  "type": "string"
                },
                {
                  "const": "-author",
                  "type": "string"
                },
                {
                  "const": "-lessonsCount",
                  "type": "string"
                },
                {
                  "const": "-enrolledParticipantsCount",
                  "type": "string"
                }
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetStudentCoursesResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/courses/available-courses": {
      "get": {
        "operationId": "CoursesController_getAvailableCourses",
        "parameters": [
          {
            "name": "title",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "category",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "author",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "creationDateRange[0]",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "creationDateRange[1]",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "state",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "archived",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "type": "number"
            }
          },
          {
            "name": "perPage",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "schema": {
              "anyOf": [
                {
                  "const": "title",
                  "type": "string"
                },
                {
                  "const": "category",
                  "type": "string"
                },
                {
                  "const": "creationDate",
                  "type": "string"
                },
                {
                  "const": "author",
                  "type": "string"
                },
                {
                  "const": "lessonsCount",
                  "type": "string"
                },
                {
                  "const": "enrolledParticipantsCount",
                  "type": "string"
                },
                {
                  "const": "-title",
                  "type": "string"
                },
                {
                  "const": "-category",
                  "type": "string"
                },
                {
                  "const": "-creationDate",
                  "type": "string"
                },
                {
                  "const": "-author",
                  "type": "string"
                },
                {
                  "const": "-lessonsCount",
                  "type": "string"
                },
                {
                  "const": "-enrolledParticipantsCount",
                  "type": "string"
                }
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetAvailableCoursesResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/courses/teacher-courses": {
      "get": {
        "operationId": "CoursesController_getTeacherCourses",
        "parameters": [
          {
            "name": "authorId",
            "required": true,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetTeacherCoursesResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/courses/course": {
      "get": {
        "operationId": "CoursesController_getCourse",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetCourseResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/courses/course-by-id": {
      "get": {
        "operationId": "CoursesController_getCourseById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetCourseByIdResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/courses/beta-course-by-id": {
      "get": {
        "operationId": "CoursesController_getBetaCourseById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetBetaCourseByIdResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/courses/{id}": {
      "patch": {
        "operationId": "CoursesController_updateCourse",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCourseBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateCourseResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/courses/enroll-course": {
      "post": {
        "operationId": "CoursesController_enrollCourse",
        "parameters": [
          {
            "name": "id",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "x-test-key",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnrollCourseResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/courses/unenroll-course": {
      "delete": {
        "operationId": "CoursesController_unenrollCourse",
        "parameters": [
          {
            "name": "id",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnenrollCourseResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "info": {
    "title": "Guidebook API",
    "description": "Example usage of Swagger with Typebox",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "GetAllCoursesResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "format": "uuid",
                  "type": "string"
                },
                "title": {
                  "type": "string"
                },
                "imageUrl": {
                  "anyOf": [
                    {
                      "type": "string"
                    },
                    {
                      "type": "null"
                    }
                  ]
                },
                "description": {
                  "type": "string"
                },
                "authorId": {
                  "format": "uuid",
                  "type": "string"
                },
                "author": {
                  "type": "string"
                },
                "authorEmail": {
                  "type": "string"
                },
                "category": {
                  "type": "string"
                },
                "courseLessonCount": {
                  "type": "number"
                },
                "completedLessonCount": {
                  "type": "number"
                },
                "enrolled": {
                  "type": "boolean"
                },
                "enrolledParticipantCount": {
                  "type": "number"
                },
                "priceInCents": {
                  "type": "number"
                },
                "currency": {
                  "type": "string"
                },
                "state": {
                  "type": "string"
                },
                "archived": {
                  "type": "boolean"
                },
                "createdAt": {
                  "type": "string"
                },
                "hasFreeLessons": {
                  "type": "boolean"
                }
              },
              "required": [
                "id",
                "title",
                "imageUrl",
                "description",
                "author",
                "category",
                "courseLessonCount",
                "completedLessonCount",
                "enrolledParticipantCount",
                "priceInCents",
                "currency"
              ]
            }
          },
          "pagination": {
            "type": "object",
            "properties": {
              "totalItems": {
                "type": "number"
              },
              "page": {
                "type": "number"
              },
              "perPage": {
                "type": "number"
              }
            },
            "required": [
              "totalItems",
              "page",
              "perPage"
            ]
          }
        },
        "required": [
          "data",
          "pagination"
        ]
      },
      "GetStudentCoursesResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "format": "uuid",
                  "type": "string"
                },
                "title": {
                  "type": "string"
                },
                "imageUrl": {
                  "anyOf": [
                    {
                      "type": "string"
                    },
                    {
                      "type": "null"
                    }
                  ]
                },
                "description": {
                  "type": "string"
                },
                "authorId": {
                  "format": "uuid",
                  "type": "string"
                },
                "author": {
                  "type": "string"
                },
                "authorEmail": {
                  "type": "string"
                },
                "category": {
                  "type": "string"
                },
                "courseLessonCount": {
                  "type": "number"
                },
                "completedLessonCount": {
                  "type": "number"
                },
                "enrolled": {
                  "type": "boolean"
                },
                "enrolledParticipantCount": {
                  "type": "number"
                },
                "priceInCents": {
                  "type": "number"
                },
                "currency": {
                  "type": "string"
                },
                "state": {
                  "type": "string"
                },
                "archived": {
                  "type": "boolean"
                },
                "createdAt": {
                  "type": "string"
                },
                "hasFreeLessons": {
                  "type": "boolean"
                }
              },
              "required": [
                "id",
                "title",
                "imageUrl",
                "description",
                "author",
                "category",
                "courseLessonCount",
                "completedLessonCount",
                "enrolledParticipantCount",
                "priceInCents",
                "currency"
              ]
            }
          },
          "pagination": {
            "type": "object",
            "properties": {
              "totalItems": {
                "type": "number"
              },
              "page": {
                "type": "number"
              },
              "perPage": {
                "type": "number"
              }
            },
            "required": [
              "totalItems",
              "page",
              "perPage"
            ]
          }
        },
        "required": [
          "data",
          "pagination"
        ]
      },
...
      "CreateCourseBody": {
        "type": "object",
        "allOf": [
          {
            "type": "object",
            "properties": {
              "title": {
                "type": "string"
              },
              "description": {
                "type": "string"
              },
              "state": {
                "anyOf": [
                  {
                    "const": "draft",
                    "type": "string"
                  },
                  {
                    "const": "published",
                    "type": "string"
                  }
                ]
              },
              "imageUrl": {
                "type": "string"
              },
              "priceInCents": {
                "type": "integer"
              },
              "currency": {
                "type": "string"
              },
              "categoryId": {
                "format": "uuid",
                "type": "string"
              }
            },
            "required": [
              "title",
              "description",
              "state",
              "categoryId"
            ]
          },
          {
            "type": "object",
            "properties": {
              "lessons": {
                "type": "array",
                "items": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          }
        ]
      },
      "CreateCourseResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object",
            "properties": {
              "id": {
                "format": "uuid",
                "type": "string"
              },
              "message": {
                "type": "string"
              }
            },
            "required": [
              "id",
              "message"
            ]
          }
        },
        "required": [
          "data"
        ]
      },
      "UpdateCourseBody": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "imageUrl": {
            "type": "string"
          },
          "state": {
            "anyOf": [
              {
                "const": "draft",
                "type": "string"
              },
              {
                "const": "published",
                "type": "string"
              }
            ]
          },
          "priceInCents": {
            "type": "integer"
          },
          "currency": {
            "type": "string"
          },
          "categoryId": {
            "format": "uuid",
            "type": "string"
          },
          "lessons": {
            "type": "array",
            "items": {
              "format": "uuid",
              "type": "string"
            }
          },
          "archived": {
            "type": "boolean"
          }
        }
      },
      "UpdateCourseResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object",
            "properties": {
              "message": {
                "type": "string"
              }
            },
            "required": [
              "message"
            ]
          }
        },
        "required": [
          "data"
        ]
      },
      "EnrollCourseResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object",
            "properties": {
              "message": {
                "type": "string"
              }
            },
            "required": [
              "message"
            ]
          }
        },
        "required": [
          "data"
        ]
      },
      "UnenrollCourseResponse": {
        "type": "null"
      },
...
    }
  }
}